TASK: Move dragndrop functionality into a single guest frame component
