+++ packages/neos-ui-guest-frame/src/InlineUI/DragAndDropUi/style.module.css
